refactor[fzs-35]: altera os diretórios e imports dos arquivos de marcações
